+++ src/storage/store/filesystem.rs
@@ -6,12 +6,11 @@ use crate::{
     array::MaybeBytes,
     byte_range::{ByteOffset, ByteRange},
     storage::{
-        ListableStorageTraits, ReadableStorageTraits, StorageError, StoreKeyRange,
-        StoreKeyStartValue, StoreKeysPrefixes, WritableStorageTraits,
+        ListableStorageTraits, ReadableStorageTraits, StorageError, StoreKey, StoreKeyError,
+        StoreKeyRange, StoreKeyStartValue, StoreKeys, StoreKeysPrefixes, StorePrefix,
+        StorePrefixes, WritableStorageTraits,
     },
 };
-
-use super::{StoreKey, StoreKeyError, StoreKeys, StorePrefix, StorePrefixes};
 
 use parking_lot::RwLock;
 use thiserror::Error;