+++ CHANGELOG.md
@@ -87,6 +87,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - **Breaking**: Partial decoders now must implement `partial_decode_opt` and the `decoded_representation` is now supplied on creation, rather than when calling `partial_decode`/`partial_decode_par`/`partial_decode_opt`
  - Sharding partial decoder now decodes inner chunks in full rather than partially decoding them, this is must faster with some codecs (e.g. blosc)
     - In future, this will probably configurable
+ - Moved `storage/store/{key.rs,prefix.rs}` to `storage/{store_key.rs,store_prefix.rs}`
 
 ### Fixed
  - Bytes codec handling of complex and raw bits data types